+++ jaguar_jwt/CHANGELOG.md
@@ -10,6 +10,7 @@
 + Tidy up for static analysis and Dart linter.
 + Implemented toString method for JwtClaim.
 + Allow for customized checking of the JWT header.
++ Fixed use of _splayify/_spaly in toJson and changed dynamic to Object.
 
 ## 2.1.2